Add nextflow params to checkatlas --with_dag --with_timeline --with_report + Cleanup files
