--- test_result/demo/Fetal.tsv
@@ -1,5 +1,0 @@
-Annot_Sample	Reference	obs	rand_index	adj_mutual_info	isolated_f1_score	rand_index_running_time	adj_mutual_info_running_time	isolated_f1_score_running_time
-Fetal_cell_type	author_cell_type	cell_type	0.9338405682824856	0.8929112773226082	0.11145855615340873	0.14680147171020508	0.2233133316040039	0.061020612716674805
-Fetal_cell_type_ontology_term_id	author_cell_type	cell_type_ontology_term_id	0.9338405682824856	0.8929112773226084	0.0	0.04853367805480957	0.08537411689758301	0.04902982711791992
-Fetal_orig.ident	author_cell_type	orig.ident	0.8458873605141566	0.19805720341469793	0.029332660728465732	0.06673049926757812	0.16042566299438477	0.06438493728637695
-Fetal_seurat_clusters	author_cell_type	seurat_clusters	0.9301595090946928	0.6885380690167049	0.0030111883562418795	0.03586435317993164	0.1828446388244629	0.06806492805480957
