--- test_result/pipeline_info/execution_trace_2026-05-18_12-12-10.txt
@@ -1,8 +1,0 @@
-task_id	hash	native_id	name	status	exit	submit	duration	realtime	%cpu	peak_rss	peak_vmem	rchar	wchar
-1	61/64b297	278629	NF_CHECKATLAS:CHECKATLAS:LIST_SCANPY_ATLASES	COMPLETED	0	2026-05-18 12:12:14.154	7.9s	7.8s	1005.4%	355.7 MB	16.3 GB	128.7 MB	14.7 KB
-7	f4/c206ec	280404	NF_CHECKATLAS:CHECKATLAS:CUSTOM_DUMPSOFTWAREVERSIONS (1)	COMPLETED	0	2026-05-18 12:12:22.189	267ms	55ms	39.6%	0	0	801.6 KB	1.7 KB
-3	ce/6546b7	280209	NF_CHECKATLAS:CHECKATLAS:CHECKATLAS_SCANPY:METRIC_DIMRED (1)	COMPLETED	0	2026-05-18 12:12:22.127	1m 24s	1m 24s	117.6%	9.4 GB	25.4 GB	2.5 GB	7.7 KB
-4	b1/59def6	280266	NF_CHECKATLAS:CHECKATLAS:CHECKATLAS_SCANPY:METRIC_CLUST (1)	COMPLETED	0	2026-05-18 12:12:22.156	1m 27s	1m 27s	122.5%	9.4 GB	25.4 GB	2.5 GB	8.1 KB
-6	9b/84ec3c	280232	NF_CHECKATLAS:CHECKATLAS:CHECKATLAS_SCANPY:SUMMARY (1)	COMPLETED	0	2026-05-18 12:12:22.141	1m 28s	1m 28s	115.0%	9.4 GB	25.4 GB	2.5 GB	316.4 KB
-5	4e/51d453	280249	NF_CHECKATLAS:CHECKATLAS:CHECKATLAS_SCANPY:QC (1)	COMPLETED	0	2026-05-18 12:12:22.148	1m 39s	1m 39s	138.2%	9.6 GB	25.6 GB	2.5 GB	4.8 MB
-2	c4/d227dc	280217	NF_CHECKATLAS:CHECKATLAS:CHECKATLAS_SCANPY:METRIC_ANNOT (1)	ABORTED	-	2026-05-18 12:12:22.135	-	-	-	-	-	-	-
