--- test_result/pipeline_info/execution_trace_2026-05-18_12-19-30.txt
@@ -1,8 +1,0 @@
-task_id	hash	native_id	name	status	exit	submit	duration	realtime	%cpu	peak_rss	peak_vmem	rchar	wchar
-1	a1/ff4302	301437	NF_CHECKATLAS:CHECKATLAS:LIST_SCANPY_ATLASES	COMPLETED	0	2026-05-18 12:19:33.106	7.8s	7.7s	989.6%	357.4 MB	16.3 GB	128.7 MB	14.7 KB
-7	65/dd8e0a	303251	NF_CHECKATLAS:CHECKATLAS:CUSTOM_DUMPSOFTWAREVERSIONS (1)	COMPLETED	0	2026-05-18 12:19:41.106	274ms	137ms	38.4%	0	0	801.6 KB	1.7 KB
-6	11/e3c17d	303087	NF_CHECKATLAS:CHECKATLAS:CHECKATLAS_SCANPY:METRIC_CLUST (1)	COMPLETED	0	2026-05-18 12:19:41.055	39.2s	39.2s	157.0%	9.2 GB	25.4 GB	2.5 GB	8.1 KB
-2	f5/15f6e1	303066	NF_CHECKATLAS:CHECKATLAS:CHECKATLAS_SCANPY:QC (1)	COMPLETED	0	2026-05-18 12:19:41.044	49s	48.9s	187.5%	9.8 GB	25.6 GB	2.5 GB	4.8 MB
-4	6b/7332ca	303076	NF_CHECKATLAS:CHECKATLAS:CHECKATLAS_SCANPY:METRIC_DIMRED (1)	COMPLETED	0	2026-05-18 12:19:41.050	56.7s	56.7s	138.0%	8.9 GB	25.4 GB	2.5 GB	7.7 KB
-5	92/9adcc0	303120	NF_CHECKATLAS:CHECKATLAS:CHECKATLAS_SCANPY:SUMMARY (1)	COMPLETED	0	2026-05-18 12:19:41.066	1m 1s	1m 1s	128.1%	9.3 GB	25.4 GB	2.5 GB	316.4 KB
-3	16/bdbcf4	303061	NF_CHECKATLAS:CHECKATLAS:CHECKATLAS_SCANPY:METRIC_ANNOT (1)	FAILED	129	2026-05-18 12:19:41.037	3h 48m 25s	3h 48m 25s	-	-	-	-	-
