--- test_result/pipeline_info/execution_trace_2026-05-19_12-13-37.txt
@@ -1,8 +1,0 @@
-task_id	hash	native_id	name	status	exit	submit	duration	realtime	%cpu	peak_rss	peak_vmem	rchar	wchar
-1	b0/df10b6	1867807	NF_CHECKATLAS:CHECKATLAS:LIST_SCANPY_ATLASES	COMPLETED	0	2026-05-19 12:13:40.573	6.5s	6.4s	1207.4%	361.9 MB	16.3 GB	128.7 MB	15.9 KB
-7	b9/8b85ea	1869487	NF_CHECKATLAS:CHECKATLAS:CUSTOM_DUMPSOFTWAREVERSIONS (1)	COMPLETED	0	2026-05-19 12:13:47.215	204ms	163ms	46.7%	0	0	801.6 KB	1.7 KB
-2	d2/f06921	1869359	NF_CHECKATLAS:CHECKATLAS:CHECKATLAS_SCANPY:METRIC_CLUST (1)	COMPLETED	0	2026-05-19 12:13:47.184	36.9s	36.8s	130.4%	8.3 GB	25.4 GB	2.5 GB	8.7 KB
-4	67/9db126	1869342	NF_CHECKATLAS:CHECKATLAS:CHECKATLAS_SCANPY:METRIC_DIMRED (1)	COMPLETED	0	2026-05-19 12:13:47.178	37s	36.8s	141.2%	8.3 GB	25.4 GB	2.5 GB	8.3 KB
-6	ae/4570a5	1869390	NF_CHECKATLAS:CHECKATLAS:CHECKATLAS_SCANPY:SUMMARY (1)	COMPLETED	0	2026-05-19 12:13:47.194	38.5s	38.5s	148.4%	9.5 GB	25.5 GB	2.5 GB	317 KB
-5	3c/c742d8	1869326	NF_CHECKATLAS:CHECKATLAS:CHECKATLAS_SCANPY:QC (1)	COMPLETED	0	2026-05-19 12:13:47.164	52.4s	52.3s	166.5%	9.6 GB	25.6 GB	2.5 GB	4.8 MB
-3	ba/6bdbd9	1869329	NF_CHECKATLAS:CHECKATLAS:CHECKATLAS_SCANPY:METRIC_ANNOT (1)	ABORTED	-	2026-05-19 12:13:47.172	-	-	-	-	-	-	-
