--- test_result/pipeline_info/pipeline_dag_2026-05-19_12-13-37.html
@@ -1,123 +1,0 @@
-<!--
-  ~ Copyright 2013-2024, Seqera Labs
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-<html>
-<head>
-<meta name="viewport" content="width=device-width, user-scalable=no, initial-scale=1, maximum-scale=1">
-</head>
-<body>
-<pre class="mermaid" style="text-align: center;">
-%%{
-  init: {
-    'theme': 'base',
-    'themeVariables': {
-      'primaryColor': '#B6ECE2',
-      'primaryTextColor': '#160F26',
-      'primaryBorderColor': '#065647',
-      'lineColor': '#545555',
-      'clusterBkg': '#BABCBD22',
-      'clusterBorder': '#DDDEDE',
-      'fontFamily': 'arial'
-    }
-  }
-}%%
-flowchart TB
-    subgraph " "
-    v0["Channel.fromPath"]
-    v4["ch_search_path"]
-    v25["ch_workflow_summary"]
-    v28["ch_methods_description"]
-    v42["checkatlas_workingdir"]
-    end
-    subgraph "NF_CHECKATLAS [NF_CHECKATLAS]"
-    subgraph "NF_CHECKATLAS:CHECKATLAS [CHECKATLAS]"
-    v5(["LIST_SCANPY_ATLASES"])
-    subgraph "NF_CHECKATLAS:CHECKATLAS:CHECKATLAS_SCANPY [CHECKATLAS_SCANPY]"
-    v8(["SUMMARY"])
-    v9(["QC"])
-    v10(["METRIC_CLUST"])
-    v11(["METRIC_ANNOT"])
-    v12(["METRIC_DIMRED"])
-    v13(( ))
-    end
-    v18(["CREATE_REPORT"])
-    v21(["CUSTOM_DUMPSOFTWAREVERSIONS"])
-    v37(["MULTIQC"])
-    v43(["COPY_MULTIQC_REPORT"])
-    v6(( ))
-    v16(( ))
-    v17(( ))
-    v26(( ))
-    v34(( ))
-    v35(( ))
-    v36(( ))
-    v40(( ))
-    end
-    end
-    subgraph " "
-    v22[" "]
-    v23[" "]
-    v38[" "]
-    v39[" "]
-    v41[" "]
-    end
-    v0 --> v34
-    v4 --> v5
-    v4 --> v8
-    v4 --> v9
-    v4 --> v10
-    v4 --> v11
-    v4 --> v12
-    v4 --> v18
-    v4 --> v37
-    v5 --> v6
-    v5 --> v16
-    v6 --> v8
-    v8 --> v13
-    v6 --> v9
-    v9 --> v13
-    v6 --> v10
-    v10 --> v13
-    v6 --> v11
-    v11 --> v13
-    v6 --> v12
-    v12 --> v13
-    v17 --> v18
-    v18 --> v37
-    v16 --> v21
-    v21 --> v23
-    v21 --> v22
-    v21 --> v26
-    v25 --> v26
-    v28 --> v26
-    v26 --> v37
-    v34 --> v37
-    v35 --> v37
-    v36 --> v37
-    v37 --> v43
-    v37 --> v39
-    v37 --> v38
-    v37 --> v40
-    v40 --> v41
-    v42 --> v43
-    v13 --> v17
-
-</pre>
-<script type="module">
-  import mermaid from 'https://cdn.jsdelivr.net/npm/mermaid@11/dist/mermaid.esm.min.mjs';
-  mermaid.initialize({ startOnLoad: true, maxTextSize: Number.MAX_SAFE_INTEGER, maxEdges: Number.MAX_SAFE_INTEGER });
-</script>
-</body>
-</html>
